pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["mahmoudmohamed57 <mahmoudmohamed572000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
Django = "^4.1.2"
django-model-utils = "^4.2.0"
django-imagekit = "^4.1.0"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.4"
django-rest-knox = "^4.2.0"
pytest-django = "^4.5.2"
django-extensions = "^3.2.1"
model-bakery = "^1.9.0"
django-filter = "^22.1"
redis = "^4.3.4"
celery = "^5.2.7"
pillow = "^9.3.0"
django-environ = "^0.9.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
